Combine type_map and return_map.py
