+++ magicgui/widgets/_bases/create_widget.py
@@ -20,9 +20,9 @@ def create_widget(
     label=None,
     gui_only=False,
     app=None,
-    is_input: bool = True,
     widget_type: str | type[_protocols.WidgetProtocol] | None = None,
     options: WidgetOptions = dict(),
+    is_result: bool = False,
 ):
     """Create and return appropriate widget subclass.
 
@@ -50,9 +50,6 @@ def create_widget(
     gui_only : bool, optional
         Whether the widget should be considered "only for the gui", or if it should
         be included in any widget container signatures, by default False
-    is_input : boolean, optional
-        Whether the widget belongs to an input or an output. By defult, an input
-        is assumed.
     app : str, optional
         The backend to use, by default ``None``
     widget_type : str or Type[WidgetProtocol] or None
@@ -62,6 +59,9 @@ def create_widget(
         autodetermined from ``value`` and/or ``annotation`` above.
     options : WidgetOptions, optional
         Dict of options to pass to the Widget constructor, by default dict()
+    is_result : boolean, optional
+        Whether the widget belongs to an input or an output. By defult, an input
+        is assumed.
 
     Returns
     -------
@@ -77,24 +77,17 @@ def create_widget(
     options = options.copy()
     kwargs = locals().copy()
     _kind = kwargs.pop("param_kind", None)
-    _is_input = kwargs.pop("is_input", None)
+    _is_result = kwargs.pop("is_result", None)
     _app = use_app(kwargs.pop("app"))
     assert _app.native
     if isinstance(widget_type, _protocols.WidgetProtocol):
         wdg_class = kwargs.pop("widget_type")
     else:
-        if _is_input:
-            from magicgui.type_map import get_widget_class as param_widget_getter
-
-            widget_getter = param_widget_getter
-        else:
-            from magicgui.return_map import get_widget_class as return_widget_getter
-
-            widget_getter = return_widget_getter
+        from magicgui.type_map import get_widget_class
 
         if widget_type:
             options["widget_type"] = widget_type
-        wdg_class, opts = widget_getter(value, annotation, options)
+        wdg_class, opts = get_widget_class(value, annotation, options, is_result) 
 
         if issubclass(wdg_class, Widget):
             opts.update(kwargs.pop("options"))
